pyproject.toml
[tool.poetry]
name = "dpo"
version = "0.1.0"
description = "Partial replication of DPO paper"
authors = ["Aaron <aaron.sandoval10@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
einops = "^0.8.0"
requests = "^2.32.3"
pytest = "^8.3.3"
tqdm = "^4.66.5"
ipykernel = "^6.29.5"
notebook = "^7.2.2"
datasets = "^3.0.1"
ipywidgets = "7.7.1"
jupyterlab-widgets = "1.1.1"
jupyter = "^1.1.1"
matplotlib = "^3.9.2"
mediapy = "^1.2.2"
pandas = "^2.2.3"
scikit-learn = "^1.5.2"
wandb = "0.13.10"
ftfy = "^6.2.3"
tensorboard = "^2.18.0"
pygame = "^2.6.1"
plotly = "^5.24.1"
sympy = "^1.13.3"
huggingface-hub = "^0.25.1"
accelerate = "^0.34.2"
pre-commit = "^3.8.0"
diffusers = "^0.30.3"
gdown = "^5.2.0"
jaxtyping = "^0.2.34"
tiktoken = "^0.7.0"
typeguard = "2.13.3"
torchinfo = "^1.8.0"
torchvision = "^0.19.1"
frozendict = "^2.4.4"
openai = "0.28"
gym = "0.23.1"
autorom = {extras = ["accept-rom-license"], version = "^0.6.1"}
ale-py = "^0.10.1"
protobuf = "3.20.3"
imageio-ffmpeg = "^0.5.1"
streamlit = "^1.38.0"
streamlit-antd-components = "0.2.5"
streamlit-image-select = "^0.6.0"
streamlit-on-hover-tabs = "^1.0.1"
nnsight = ">=0.3.5"
opencv-python = "^4.10.0.84"
torchtext = "^0.18.0"
tabulate = "^0.9.0"
safetensors = "^0.4.5"
portalocker = ">=2.0.0"
numpy = "^1.26"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
